Move initializeEdges
